Feat : 알림 기능 추가

+++ src/main/java/com/example/crudpersional/domain/entity/alarm/Alarm.java
@@ -17,10 +17,11 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 public class Alarm {
-    private Integer id;
+
+    private Long id;
     private AlarmType alarmType;
-    private Integer fromUserId;
-    private Integer targetId;
+    private Long fromUserId;
+    private Long targetId;
     private String createdAt;
